NVTK/requirements.txt
numpy>=1.15.4
scikitlearn>=0.20.0
pygame>=1.9.4
pyserial>=3.4
mne>=0.17.0
openbci-python>=1.0.0
yapsy>=1.12.0